# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Synapse Workspace API",
    "version": "0.1.0",
    "description": "The official REST and AI API specification for Synapse Knowledge & Work OS."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "System Health & Diagnostic Status",
        "description": "Returns uptime, operational mode (Local / Cloud), and active subsystem states.",
        "responses": {
          "200": {
            "description": "System is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/summarize": {
      "post": {
        "summary": "AI Note Summarizer",
        "description": "Generates concise bulleted summaries of note contents.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "example": "Long note content..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string"
                    },
                    "provider": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/expand": {
      "post": {
        "summary": "AI Thought Expander",
        "description": "Expands short bullets or rough ideas into detailed structured documentation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bullet"
                ],
                "properties": {
                  "bullet": {
                    "type": "string",
                    "example": "Local-first architecture"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expanded content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string"
                    },
                    "provider": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/generate-template": {
      "post": {
        "summary": "AI Custom Template Generator",
        "description": "Generates structured markdown blueprints for meeting notes, RFCs, sprints, and user stories.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "example": "Sprint Planning Document"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated blueprint content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "string"
                    },
                    "provider": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}